requirements.txt
Django==4.0.5
django-widget-tweaks==1.4.12
python-decouple==3.6
psycopg2==2.9.3
django-nested-admin==3.4.0
